# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API untuk sistem E-Catalog Tour \u0026 Travel dengan WhatsApp Gateway",
        "title": "Pintour E-Catalog API",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/packages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-packages"
                ],
                "summary": "Daftar paket (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kategori",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Halaman",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-packages"
                ],
                "summary": "Buat paket baru",
                "parameters": [
                    {
                        "description": "Data paket",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_irfan-ghzl_pintour-travel_internal_domain_package.Package"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/packages/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "admin-packages"
                ],
                "summary": "Update paket",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Package ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "admin-packages"
                ],
                "summary": "Hapus paket (soft delete)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Package ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/packages": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catalog"
                ],
                "summary": "Daftar paket wisata publik",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kategori (reguler/umroh/halal/honeymoon)",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Destinasi",
                        "name": "destination",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Harga minimum",
                        "name": "price_min",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Harga maksimum",
                        "name": "price_max",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Halaman",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Per halaman",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/packages/{slug}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catalog"
                ],
                "summary": "Detail paket berdasarkan slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slug paket",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_irfan-ghzl_pintour-travel_internal_domain_package.Package": {
            "type": "object",
            "required": [
                "base_price",
                "destination",
                "duration_days",
                "name"
            ],
            "properties": {
                "base_price": {
                    "type": "number"
                },
                "category": {
                    "description": "default reguler",
                    "type": "string",
                    "enum": [
                        "reguler",
                        "umroh",
                        "halal",
                        "honeymoon"
                    ]
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "duration_days": {
                    "type": "integer",
                    "maximum": 365,
                    "minimum": 1
                },
                "facilities": {
                    "description": "FR-CAT-03",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "itinerary": {
                    "description": "[]ItineraryDay",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "name": {
                    "type": "string"
                },
                "requirements": {
                    "description": "[]string",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "slug": {
                    "type": "string"
                },
                "terms_conditions": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "visa_info": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}